# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handlers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handlers.oidcConfigResponse:
    properties:
      client_id:
        type: string
      client_secret_set:
        type: boolean
      issuer_url:
        type: string
      redirect_uri:
        type: string
    type: object
...
  internal_handlers.setOIDCConfigRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      issuer_url:
        type: string
      redirect_uri:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal library management API
  title: Biblioteka API
  version: "1.0"
paths:
...
  /auth/oidc/callback:
    get:
      description: Handles the OIDC provider's redirect after authentication
      parameters:
      - description: OIDC state parameter
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Redirect to frontend with token
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      summary: OIDC callback
      tags:
      - OIDC
...
  /auth/oidc/link:
    get:
      description: Redirects to OIDC provider to link account (requires nonce)
      parameters:
      - description: Link nonce from CreateLinkNonce
        in: query
        name: nonce
        required: true
        type: string
      responses:
        "302":
          description: Redirect to OIDC provider
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      summary: Link OIDC account
      tags:
      - OIDC
  /auth/oidc/link-nonce:
    post:
      description: Generate a short-lived nonce for linking an OIDC account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              nonce:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create OIDC link nonce
      tags:
      - OIDC
  /auth/oidc/login:
    get:
      description: Redirects to the OIDC provider's authorization endpoint
      responses:
        "302":
          description: Redirect to OIDC provider
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      summary: OIDC login
      tags:
      - OIDC
...
  /config/oidc:
    get:
      description: Returns current OIDC configuration (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.oidcConfigResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get OIDC configuration
      tags:
      - Config
    put:
      consumes:
      - application/json
      description: Update OIDC configuration with validation (admin only)
      parameters:
      - description: OIDC configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.setOIDCConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Set OIDC configuration
      tags:
      - Config
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"